feat(antlr): fix some code offences
