+++ src/main/java/com/github/lombrozo/jsmith/antlr/rules/LexerRuleSpec.java
@@ -24,9 +24,8 @@
 package com.github.lombrozo.jsmith.antlr.rules;
 
 import com.github.lombrozo.jsmith.antlr.Context;
-import com.github.lombrozo.jsmith.antlr.view.RuleTextNode;
+import com.github.lombrozo.jsmith.antlr.view.RuleText;
 import com.github.lombrozo.jsmith.antlr.view.Text;
-import com.github.lombrozo.jsmith.antlr.view.TextNode;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -84,7 +83,7 @@ public final class LexerRuleSpec implements Rule {
 
     @Override
     public Text generate(final Context context) {
-        return new RuleTextNode(
+        return new RuleText(
             this,
             this.list.stream()
                 .map(rule -> rule.generate(context))
