+++ src/main/java/com/github/lombrozo/jsmith/antlr/rules/ParserRuleSpec.java
@@ -24,9 +24,8 @@
 package com.github.lombrozo.jsmith.antlr.rules;
 
 import com.github.lombrozo.jsmith.antlr.Context;
-import com.github.lombrozo.jsmith.antlr.view.RuleTextNode;
+import com.github.lombrozo.jsmith.antlr.view.RuleText;
 import com.github.lombrozo.jsmith.antlr.view.Text;
-import com.github.lombrozo.jsmith.antlr.view.TextNode;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -77,7 +76,7 @@ public final class ParserRuleSpec implements Rule {
 
     @Override
     public Text generate(final Context context) {
-        return new RuleTextNode(
+        return new RuleText(
             this,
             this.list.stream().map(rule -> rule.generate(context)).collect(Collectors.toList())
         );
